refactor : mixpanel infrastructure 단으로 이동
