--- src/main/kotlin/com/deepromeet/atcha/mixpanel/MixpanelEvent.kt
@@ -1,9 +1,0 @@
-package com.deepromeet.atcha.mixpanel
-
-import com.deepromeet.atcha.mixpanel.event.MixpanelEventProperty
-
-class MixpanelEvent(
-    val mixpanelEventName: MixpanelEventName,
-    val distinctId: String,
-    val property: MixpanelEventProperty
-)
