--- src/main/kotlin/com/deepromeet/atcha/mixpanel/event/MixpanelEventProperty.kt
@@ -1,3 +1,0 @@
-package com.deepromeet.atcha.mixpanel.event
-
-interface MixpanelEventProperty
