+++ src/main/kotlin/com/deepromeet/atcha/shared/infrastructure/mixpanel/MixpanelEvent.kt
@@ -1,0 +1,9 @@
+package com.deepromeet.atcha.shared.infrastructure.mixpanel
+
+import com.deepromeet.atcha.shared.infrastructure.mixpanel.event.MixpanelEventProperty
+
+class MixpanelEvent(
+    val mixpanelEventName: MixpanelEventName,
+    val distinctId: String,
+    val property: MixpanelEventProperty
+)
