+++ src/main/kotlin/com/deepromeet/atcha/shared/infrastructure/mixpanel/event/MixpanelEventProperty.kt
@@ -1,0 +1,3 @@
+package com.deepromeet.atcha.shared.infrastructure.mixpanel.event
+
+interface MixpanelEventProperty
